spaces/multimodal/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
jinja2>=3.1.0
python-multipart>=0.0.12
pydantic>=2.0.0
pillow>=10.0.0
torch>=2.4.0
transformers>=4.45.0
accelerate>=0.34.0
spaces>=0.30.0
qwen-vl-utils>=0.0.8
